Shopping cart implemented

+++ PharMedTOGO.Core/Models/SaleFormModel.cs
@@ -10,7 +10,7 @@ namespace PharMedTOGO.Core.Models
         [Range(SaleRangePriceMinValue, 
             SaleRangeMaxValue, 
             ErrorMessage = SaleRangeErrorMessage)]
-        public decimal Discount { get; set; }
+        public int Discount { get; set; }
 
         [Required(ErrorMessage = RequiredMessage)]
         [DataType(DataType.Date)]
